pyproject.toml
[tool.poetry]
name = "red-core"
version = "3.0.0"
description = "Comprehensive framework for adversarial AI safety research, focusing on refusal robustness testing, guardrail decay analysis, and systematic attack pattern discovery."
authors = ["Cassius Oldenburg <connect@cassius.red>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.0"
pydantic = "^2.11.5"
tiktoken = "^0.9.0"
openai = "^1.82.0"
anthropic = "^0.52.0"
google-generativeai = "^0.8.5"
rich = "^14.0.0"
tqdm = "^4.66.0"
scipy = "^1.11.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
mypy = "^1.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.group.local.dependencies]
llama-cpp-python = "^0.2.0"

[tool.poetry.extras]
local = ["llama-cpp-python"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.pytest_asyncio]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"